put i love potato controller in controllers package
